Document the new API for userscripts
Add a tip on the tips page that makes users aware of the API. The tip
includes a link to github where they can view the details.

Update userscripts/README.md to note the capabilities of the API
and link to the API js file, where developers can see function
definitions.